requirements.txt
aiofiles==23.2.1
aiogram==3.12.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
frozenlist==1.4.1
idna==3.8
joblib==1.4.2
magic-filter==1.0.12
multidict==6.0.5
nltk==3.9.1
numpy==2.1.0
pandas==2.2.2
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
regex==2024.7.24
six==1.16.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
yarl==1.9.4
